przygotowanie do szarzy

+++ src/objects/unit/Unit.h
@@ -126,5 +126,5 @@ private:
 	short posInFormation = -1;
 	short formation = -1;
 
-	int currentFrameState = 0;
+	unsigned short currentFrameState = 0;
 };
